Deconstruction of background overlays
