+++ clientd3d/d3drender.c
@@ -167,8 +167,6 @@ void				D3DRenderLMapsBuild(void);
 void				D3DLMapsStaticGet(room_type *room);
 void				D3DRenderFontInit(font_3d *pFont, HFONT hFont);
 
-void				D3DRenderBackgroundOverlays(d3d_render_pool_new* pPool, int angleHeading, int anglePitch, room_type* room, Draw3DParams* params);
-void				D3DRenderSunDraw(int angleHeading, int anglePitch);
 
 // new render stuff
 void					D3DRenderPoolInit(d3d_render_pool_new *pPool, int size, int packetSize);
@@ -580,27 +578,7 @@ void D3DRenderBegin(room_type *room, Draw3DParams *params)
 	// background overlays (e.g. the Sun & Moon)
 	if (draw_background_overlays)
 	{
-		MatrixIdentity(&mat);
-		IDirect3DDevice9_SetTransform(gpD3DDevice, D3DTS_WORLD, &mat);
-		IDirect3DDevice9_SetVertexShader(gpD3DDevice, NULL);
-		IDirect3DDevice9_SetVertexDeclaration(gpD3DDevice, decl1dc);
-
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ZWRITEENABLE, FALSE);
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ZENABLE, TRUE);
-
-		D3DRenderPoolReset(&gWorldPool, &D3DMaterialWorldPool);
-		D3DCacheSystemReset(&gWorldCacheSystem);
 		D3DRenderBackgroundOverlays(&gWorldPool, angleHeading, anglePitch, room, params);
-		D3DCacheFill(&gWorldCacheSystem, &gWorldPool, 1);
-		D3DCacheFlush(&gWorldCacheSystem, &gWorldPool, 1, D3DPT_TRIANGLESTRIP);
-
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ZWRITEENABLE, TRUE);
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_FOGENABLE, TRUE);
-
-		// restore the correct material and view matrices.
-		MatrixIdentity(&mat);
-		IDirect3DDevice9_SetTransform(gpD3DDevice, D3DTS_WORLD, &mat);
-		IDirect3DDevice9_SetTransform(gpD3DDevice, D3DTS_VIEW, &view);
 	}
 
 	if (draw_world)
@@ -1376,1209 +1354,6 @@ void D3DRenderFontInit(font_3d *pFont, HFONT hFont)
    DeleteObject(hScaledFont);
    DeleteDC(hDC);
 }
-
-void D3DRenderBackgroundOverlays(d3d_render_pool_new* pPool, int angleHeading, int anglePitch, room_type* room, Draw3DParams* params)
-{
-	for (list_type list = room->bg_overlays; list != NULL; list = list->next)
-	{
-		BackgroundOverlay* overlay = (BackgroundOverlay*)(list->data);
-
-		PDIB pDib = GetObjectPdib(overlay->obj.icon_res, 0, overlay->obj.animate->group);
-		if (NULL == pDib)
-			continue;
-
-		BYTE* bkgnd_bmap = DibPtr(pDib);
-		if (bkgnd_bmap == NULL)
-			continue;
-
-		// The background overlay is not yet considered visible to the player.
-		overlay->drawn = FALSE;
-
-		// Increase the size of the background overlay if necessary.
-		float size_scaler = 1.2f;
-
-		// Specify the maximum and minimum altitude of the background overlay.
-		// This will map from the -200 to 200 values return from the server to these values.
-		int height_max = 200;
-		int height_min = -200;
-
-		long object_width = DibWidth(pDib) * size_scaler;
-		long object_height = DibHeight(pDib) * size_scaler;
-
-		d3d_render_packet_new* pPacket = D3DRenderPacketFindMatch(pPool, NULL, pDib, 0, 0, 0);
-		if (NULL == pPacket)
-			return;
-
-		d3d_render_chunk_new* pChunk = D3DRenderChunkNew(pPacket);
-		assert(pChunk);
-
-		pPacket->pMaterialFctn = &D3DMaterialWorldPacket;
-
-		pChunk->numIndices = 4;
-		pChunk->numVertices = 4;
-		pChunk->numPrimitives = pChunk->numVertices - 2;
-		pChunk->pMaterialFctn = &D3DMaterialWorldDynamicChunk;
-		pChunk->flags |= D3DRENDER_NOAMBIENT | D3DRENDER_WORLD_OBJ;
-		pChunk->zBias = ZBIAS_BASE;
-
-		int piAngle = overlay->x;
-		int piHeight = overlay->y;
-
-		auto mapRange = [](double value, double in_min, double in_max, double out_min, double out_max) -> double {
-			return (value - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
-		};
-
-		long mappedHeightValue = mapRange(piHeight, -200, 200, height_min, height_max);
-		float angleInRadians = piAngle / 4096.0f;
-		double azimuthalAngle = angleInRadians * 2 * PI;
-		long radius = height_max * 5;
-		double horizontalRadius = sqrt(pow(radius, 2) - pow(mappedHeightValue, 2));
-
-		long x = params->viewer_x - (object_width / 2) + horizontalRadius * cos(azimuthalAngle);
-		long y = params->viewer_y - (object_height / 2) + horizontalRadius * sin(azimuthalAngle);
-		long z = params->viewer_height + mappedHeightValue;
-
-		Vector3D bg_overlay_pos;
-		bg_overlay_pos.x = x;
-		bg_overlay_pos.y = y;
-		bg_overlay_pos.z = z;
-
-		D3DMATRIX rot, mat;
-		MatrixIdentity(&mat);
-		MatrixIdentity(&rot);
-
-		const float FULL_CIRCLE_TO_DEGREES = 360.0f / 4096.0f;
-		const float DEGREES_TO_RADIANS = PI / 180.0f;
-		const float ANGLE_RANGE_TO_DEGREES = 45.0f / 414.0f;
-
-		MatrixRotateY(&rot, (float)angleHeading * FULL_CIRCLE_TO_DEGREES * DEGREES_TO_RADIANS);
-		MatrixRotateX(&mat, (float)anglePitch * ANGLE_RANGE_TO_DEGREES * DEGREES_TO_RADIANS);
-		MatrixTranspose(&rot, &rot);
-		MatrixTranslate(&mat, bg_overlay_pos.x, bg_overlay_pos.z, bg_overlay_pos.y);
-		MatrixMultiply(&pChunk->xForm, &rot, &mat);
-
-		pChunk->xyz[0].x = 0;
-		pChunk->xyz[0].z = 0;
-		pChunk->xyz[0].y = 0;
-
-		pChunk->xyz[1].x = -object_width;
-		pChunk->xyz[1].z = 0;
-		pChunk->xyz[1].y = 0;
-
-		pChunk->xyz[2].x = -object_width;
-		pChunk->xyz[2].z = object_height;
-		pChunk->xyz[2].y = 0;
-
-		pChunk->xyz[3].x = 0;
-		pChunk->xyz[3].z = object_height;
-		pChunk->xyz[3].y = 0;
-
-		for (int j = 0; j < 4; j++) {
-			pChunk->bgra[j].b = 255;
-			pChunk->bgra[j].g = 255;
-			pChunk->bgra[j].r = 255;
-			pChunk->bgra[j].a = 255;
-		}
-
-		static bool initialized = false;
-		static float u[4] = { 0.0f, 0.0f, 1.0f, 1.0f };
-		static float v[4] = { 0.0f, 1.0f, 1.0f, 0.0f };
-		const float epsilon = 0.007f; // Small value to adjust UV coordinates inward
-		if (!initialized)
-		{
-			// Set the texture coordinate template with epsilon adjustment once.
-			u[0] = epsilon;
-			v[0] = epsilon;
-			u[1] = 1.0f - epsilon;
-			v[1] = epsilon;
-			u[2] = 1.0f - epsilon;
-			v[2] = 1.0f - epsilon;
-			u[3] = epsilon;
-			v[3] = 1.0f - epsilon;
-			initialized = true;
-		}
-
-		pChunk->st0[0].s = u[0];
-		pChunk->st0[0].t = v[0];
-		pChunk->st0[1].s = u[1];
-		pChunk->st0[1].t = v[1];
-		pChunk->st0[2].s = u[2];
-		pChunk->st0[2].t = v[2];
-		pChunk->st0[3].s = u[3];
-		pChunk->st0[3].t = v[3];
-
-		pChunk->indices[0] = 1;
-		pChunk->indices[1] = 2;
-		pChunk->indices[2] = 0;
-		pChunk->indices[3] = 3;
-
-		// Determine if the background overlay is visible for click detection in client
-		// (e.g. Blinded by the light of the Sun).
-		ObjectRange* range = FindVisibleObjectById(overlay->obj.id);
-
-		int w = gD3DRect.right - gD3DRect.left;
-		int h = gD3DRect.bottom - gD3DRect.top;
-
-		custom_xyzw topLeft;
-		topLeft.x = pChunk->xyz[3].x;
-		topLeft.y = pChunk->xyz[3].z;
-		topLeft.z = 0;
-		topLeft.w = 1.0f;
-
-		custom_xyzw topRight;
-		topRight.x = pChunk->xyz[3].x;
-		topRight.y = pChunk->xyz[3].z;
-		topRight.z = 0;
-		topRight.w = 1.0f;
-
-		custom_xyzw bottomLeft;
-		bottomLeft.x = pChunk->xyz[1].x;
-		bottomLeft.y = pChunk->xyz[1].z;
-		bottomLeft.z = 0;
-		bottomLeft.w = 1.0f;
-
-		custom_xyzw bottomRight;
-		bottomRight.x = pChunk->xyz[1].x;
-		bottomRight.y = pChunk->xyz[1].z;
-		bottomRight.z = 0;
-		bottomRight.w = 1.0f;
-
-		D3DMATRIX localToScreen, trans;
-		MatrixRotateY(&rot, (float)angleHeading * 360.0f / 4096.0f * PI / 180.0f);
-		MatrixRotateX(&mat, (float)anglePitch * 50.0f / 414.0f * PI / 180.0f);
-		MatrixMultiply(&rot, &rot, &mat);
-		MatrixTranslate(&trans, -(float)params->viewer_x, -(float)params->viewer_height, -(float)params->viewer_y);
-		MatrixMultiply(&mat, &trans, &rot);
-		XformMatrixPerspective(&localToScreen, FovHorizontal(gD3DRect.right - gD3DRect.left), FovVertical(gD3DRect.bottom - gD3DRect.top), 1.0f, 2000000.0f);
-		MatrixMultiply(&mat, &pChunk->xForm, &mat);
-		MatrixMultiply(&localToScreen, &mat, &localToScreen);
-
-		MatrixMultiplyVector(&topLeft, &localToScreen, &topLeft);
-		MatrixMultiplyVector(&topRight, &localToScreen, &topRight);
-		MatrixMultiplyVector(&bottomLeft, &localToScreen, &bottomLeft);
-		MatrixMultiplyVector(&bottomRight, &localToScreen, &bottomRight);
-
-		topLeft.x /= topLeft.w;
-		topLeft.y /= topLeft.w;
-		topLeft.z /= topLeft.w;
-		bottomRight.x /= bottomRight.w;
-		bottomRight.y /= bottomRight.w;
-		bottomRight.z /= bottomRight.w;
-
-		topLeft.z = topLeft.z * 2.0f - 1.0f;
-		bottomRight.z = bottomRight.z * 2.0f - 1.0f;
-
-		topRight.x = bottomRight.x;
-		topRight.y = topLeft.y;
-		topRight.z = topLeft.z;
-		bottomLeft.x = topLeft.x;
-		bottomLeft.y = bottomRight.y;
-		bottomLeft.z = topLeft.z;
-
-		custom_xyzw center;
-		center.x = (topLeft.x + topRight.x) / 2.0f;
-		center.y = (topLeft.y + bottomLeft.y) / 2.0f;
-		center.z = topLeft.z;
-
-		if (
-			(
-				(D3DRENDER_CLIP(topLeft.x, 1.0f) &&
-					D3DRENDER_CLIP(topLeft.y, 1.0f)) ||
-				(D3DRENDER_CLIP(bottomLeft.x, 1.0f) &&
-					D3DRENDER_CLIP(bottomLeft.y, 1.0f)) ||
-				(D3DRENDER_CLIP(topRight.x, 1.0f) &&
-					D3DRENDER_CLIP(topRight.y, 1.0f)) ||
-				(D3DRENDER_CLIP(bottomRight.x, 1.0f) &&
-					D3DRENDER_CLIP(bottomRight.y, 1.0f)) ||
-				(D3DRENDER_CLIP(center.x, 1.0f))
-				) &&
-			D3DRENDER_CLIP(topLeft.z, 1.0f))
-		{
-
-			int tempLeft = (topLeft.x * w / 2) + (w / 2);
-			int tempRight = (bottomRight.x * w / 2) + (w / 2);
-			int tempTop = (topLeft.y * -h / 2) + (h / 2);
-			int tempBottom = (bottomRight.y * -h / 2) + (h / 2);
-
-			int distX = bg_overlay_pos.x - player.x;
-			int distY = bg_overlay_pos.y - player.y;
-
-			int distance = DistanceGet(distX, distY);
-
-			if (range == NULL)
-			{
-				// Set up new visible object.
-				range = &visible_objects[num_visible_objects];
-				range->id = overlay->obj.id;
-				range->distance = distance;
-				range->left_col = tempLeft;
-				range->right_col = tempRight;
-				range->top_row = tempTop;
-				range->bottom_row = tempBottom;
-
-				num_visible_objects = min(num_visible_objects + 1, MAXOBJECTS);
-			}
-
-			overlay->rcScreen.left = tempLeft;
-			overlay->rcScreen.right = tempRight;
-			overlay->rcScreen.top = tempTop;
-			overlay->rcScreen.bottom = tempBottom;
-
-			// The background overlay is visible and eligable for click detection.
-			overlay->drawn = TRUE;
-
-			// Record boundaries of drawing area.
-			range->left_col = min(range->left_col, tempLeft);
-			range->right_col = max(range->right_col, tempRight);
-			range->top_row = min(range->top_row, tempTop);
-			range->bottom_row = max(range->bottom_row, tempBottom);
-		}
-
-	}
-}
-
-int D3DRenderWallExtract(WallData *pWall, PDIB pDib, unsigned int *flags, custom_xyz *pXYZ,
-						  custom_st *pST, custom_bgra *pBGRA, unsigned int type, int side)
-{
-	int				top, bottom;
-	int				xOffset, yOffset;
-	Sidedef			*pSideDef;
-	int				drawTopDown;
-	int				paletteIndex;
-	BYTE			light;
-
-	if (pXYZ)
-	{
-		// pos and neg sidedefs have their x and y coords reversed
-		if (side > 0)
-		{
-			pSideDef = pWall->pos_sidedef;
-
-			if (NULL == pWall->pos_sector)
-				light = 0;
-			else
-				light = pWall->pos_sector->light;
-
-			xOffset = pWall->pos_xoffset;
-			yOffset = pWall->pos_yoffset;
-
-			pXYZ[0].x = pWall->x0;
-			pXYZ[3].x = pWall->x1;
-			pXYZ[1].x = pWall->x0;
-			pXYZ[2].x = pWall->x1;
-
-			pXYZ[0].y = pWall->y0;
-			pXYZ[3].y = pWall->y1;
-			pXYZ[1].y = pWall->y0;
-			pXYZ[2].y = pWall->y1;
-
-			switch (type)
-			{
-				case D3DRENDER_WALL_NORMAL:
-				{
-					pXYZ[0].z = pWall->z2;
-					pXYZ[3].z = pWall->zz2;
-					pXYZ[1].z = pWall->z1;
-					pXYZ[2].z = pWall->zz1;
-
-					if (pSideDef->flags & WF_NORMAL_TOPDOWN)
-						drawTopDown = 1;
-					else
-						drawTopDown = 0;
-				}
-				break;
-
-				case D3DRENDER_WALL_BELOW:
-				{
-					if ((pWall->bowtie_bits & BT_BELOW_POS) ||
-						(pWall->bowtie_bits & BT_BELOW_NEG))
-					{
-						pXYZ[0].z = pWall->z1Neg;
-						pXYZ[3].z = pWall->zz1Neg;
-					}
-					else
-					{
-						pXYZ[0].z = pWall->z1;
-						pXYZ[3].z = pWall->zz1;
-					}
-
-					pXYZ[1].z = pWall->z0;
-					pXYZ[2].z = pWall->zz0;
-
-					if (pSideDef->flags & WF_BELOW_TOPDOWN)
-						drawTopDown = 1;
-					else
-						drawTopDown = 0;
-				}
-				break;
-
-				case D3DRENDER_WALL_ABOVE:
-				{
-					pXYZ[0].z = pWall->z3;
-					pXYZ[3].z = pWall->zz3;
-					pXYZ[1].z = pWall->z2;
-					pXYZ[2].z = pWall->zz2;
-
-					if (pSideDef->flags & WF_ABOVE_BOTTOMUP)
-						drawTopDown = 0;
-					else
-						drawTopDown = 1;
-				}
-				break;
-
-				default:
-				break;
-			}
-		}
-		else if (side < 0)
-		{
-			pSideDef = pWall->neg_sidedef;
-
-			if (NULL == pWall->neg_sector)
-				light = 0;
-			else
-				light = pWall->neg_sector->light;
-
-			xOffset = pWall->neg_xoffset;
-			yOffset = pWall->neg_yoffset;
-
-			pXYZ[0].x = (float)pWall->x1;
-			pXYZ[3].x = (float)pWall->x0;
-			pXYZ[1].x = (float)pWall->x1;
-			pXYZ[2].x = (float)pWall->x0;
-
-			pXYZ[0].y = (float)pWall->y1;
-			pXYZ[3].y = (float)pWall->y0;
-			pXYZ[1].y = (float)pWall->y1;
-			pXYZ[2].y = (float)pWall->y0;
-
-			switch (type)
-			{
-				case D3DRENDER_WALL_NORMAL:
-				{
-					pXYZ[0].z = (long)pWall->zz2;
-					pXYZ[3].z = (long)pWall->z2;
-					pXYZ[1].z = (long)pWall->zz1;
-					pXYZ[2].z = (long)pWall->z1;
-
-					if (pSideDef->flags & WF_NORMAL_TOPDOWN)
-						drawTopDown = 1;
-					else
-						drawTopDown = 0;
-				}
-				break;
-
-				case D3DRENDER_WALL_BELOW:
-				{
-					pXYZ[0].z = (long)pWall->zz1;
-					pXYZ[3].z = (long)pWall->z1;
-					pXYZ[1].z = (long)pWall->zz0;
-					pXYZ[2].z = (long)pWall->z0;
-
-					if (pSideDef->flags & WF_BELOW_TOPDOWN)
-						drawTopDown = 1;
-					else
-						drawTopDown = 0;
-				}
-				break;
-
-				case D3DRENDER_WALL_ABOVE:
-				{
-					pXYZ[0].z = (long)pWall->zz3;
-					pXYZ[3].z = (long)pWall->z3;
-					pXYZ[1].z = (long)pWall->zz2;
-					pXYZ[2].z = (long)pWall->z2;
-
-					if (pSideDef->flags & WF_ABOVE_BOTTOMUP)
-						drawTopDown = 0;
-					else
-						drawTopDown = 1;
-				}
-				break;
-
-				default:
-				break;
-			}
-		}
-		else
-			assert(0);
-	}
-
-	*flags = 0;
-
-	if (pSideDef->flags & WF_TRANSPARENT)
-		*flags |= D3DRENDER_TRANSPARENT;
-
-	switch (type)
-	{
-		case D3DRENDER_WALL_NORMAL:
-			if (pSideDef->flags & WF_NO_VTILE)
-				*flags |= D3DRENDER_NO_VTILE;
-			if (pSideDef->flags & WF_NO_HTILE)
-				*flags |= D3DRENDER_NO_HTILE;
-		break;
-
-		default:
-      break;
-	}
-
-	if ((pXYZ) && (pST))
-	{
-		float	invWidth, invHeight, invWidthFudge, invHeightFudge;
-
-		// force a wraparound because many textures in the old client do this, grr
-		yOffset = yOffset << 16;
-		yOffset = yOffset >> 16;
-
-		invWidth = 1.0f / (float)pDib->width;
-		invHeight = 1.0f / (float)pDib->height;
-		invWidthFudge = 1.0f / ((float)pDib->width * PETER_FUDGE);
-		invHeightFudge = 1.0f / ((float)pDib->height * PETER_FUDGE);
-
-		pST[0].s = (float)xOffset * (float)(pDib->shrink) * invHeight;
-		pST[1].s = (float)xOffset * (float)(pDib->shrink) * invHeight;
-		pST[3].s = (float)(pST[0].s + ((float)pWall->length * (float)pDib->shrink) * invHeight);
-		pST[2].s = (float)(pST[1].s + ((float)pWall->length * (float)pDib->shrink) * invHeight);
-
-		if (!drawTopDown)
-		{
-			if (pXYZ[1].z == pXYZ[2].z)
-				bottom = pXYZ[1].z;
-			else
-			{
-				bottom = min(pXYZ[1].z, pXYZ[2].z);
-				bottom = bottom & ~(FINENESS - 1);
-			}
-
-			if (pXYZ[0].z == pXYZ[3].z)
-				top = pXYZ[0].z;
-			else
-			{
-				top = max(pXYZ[0].z, pXYZ[3].z);
-				top = (top + FINENESS - 1) & ~(FINENESS - 1);
-			}
-
-			if (pXYZ[1].z == pXYZ[2].z)
-			{
-				pST[1].t = 1.0f - (((float)yOffset * (float)pDib->shrink)
-					* invWidth);
-				pST[2].t = 1.0f - (((float)yOffset * (float)pDib->shrink)
-					* invWidth);
-			}
-			else
-			{
-				pST[1].t = 1.0f - (((float)yOffset * (float)pDib->shrink)
-					*invWidth);
-				pST[2].t = 1.0f - (((float)yOffset * (float)pDib->shrink)
-					* invWidth);
-				pST[1].t -= ((float)pXYZ[1].z - bottom) * (float)pDib->shrink
-						* invWidthFudge;
-				pST[2].t -= ((float)pXYZ[2].z - bottom) * (float)pDib->shrink
-						* invWidthFudge;
-			}
-
-			pST[0].t = pST[1].t -
-				((float)(pXYZ[0].z - pXYZ[1].z) * (float)pDib->shrink * invWidthFudge);
-			pST[3].t = pST[2].t -
-				((float)(pXYZ[3].z - pXYZ[2].z) * (float)pDib->shrink * invWidthFudge);
-		}
-		else	// else, need to place tex origin at top left
-		{
-			if (pXYZ[0].z == pXYZ[3].z)
-				top = pXYZ[0].z;
-			else
-			{
-				top = max(pXYZ[0].z, pXYZ[3].z);
-				top = (top + FINENESS - 1) & ~(FINENESS - 1);
-			}
-
-			if (pXYZ[1].z == pXYZ[2].z)
-				bottom = pXYZ[1].z;
-			else
-			{
-				bottom = min(pXYZ[1].z, pXYZ[2].z);
-				bottom = bottom & ~(FINENESS - 1);
-			}
-
-			if (pXYZ[0].z == pXYZ[3].z)
-			{
-				pST[0].t = 0.0f;
-				pST[3].t = 0.0f;
-			}
-			else
-			{
-				pST[0].t = ((float)top - pXYZ[0].z) * (float)pDib->shrink
-						* invWidthFudge;
-				pST[3].t = ((float)top - pXYZ[3].z) * (float)pDib->shrink
-						* invWidthFudge;
-			}
-
-			pST[0].t -= ((float)(yOffset * pDib->shrink) * invWidth);
-			pST[3].t -= ((float)(yOffset * pDib->shrink) * invWidth);
-
-			pST[1].t = pST[0].t + ((pXYZ[0].z - pXYZ[1].z) * (float)pDib->shrink
-				* invWidthFudge);
-			pST[2].t = pST[3].t + ((pXYZ[3].z - pXYZ[2].z) * (float)pDib->shrink
-				* invWidthFudge);
-		}
-
-		if (pSideDef->animate != NULL && pSideDef->animate->animation == ANIMATE_SCROLL)
-		{
-			int	i;
-			if (pSideDef->flags & WF_BACKWARDS)
-			{
-				for (i = 0; i < 4; i++)
-				{
-					pST[i].s -= pSideDef->animate->u.scroll.xoffset * pDib->shrink *
-						invHeight;
-					pST[i].t += pSideDef->animate->u.scroll.yoffset * pDib->shrink *
-						invWidth;
-				}
-			}
-			else
-			{
-				for (i = 0; i < 4; i++)
-				{
-					pST[i].s += pSideDef->animate->u.scroll.xoffset * pDib->shrink *
-						invHeight;
-					pST[i].t -= pSideDef->animate->u.scroll.yoffset * pDib->shrink *
-						invWidth;
-				}
-			}
-		}
-
-		if (pSideDef->flags & WF_BACKWARDS)
-		{
-			float	temp;
-
-			temp = pST[3].s;
-			pST[3].s = pST[0].s;
-			pST[0].s = temp;
-
-			temp = pST[2].s;
-			pST[2].s = pST[1].s;
-			pST[1].s = temp;
-		}
-
-		if (*flags & D3DRENDER_NO_VTILE)
-		{
-			if (pST[0].t < 0.0f)
-			{
-				float	tex, wall, ratio, temp;
-
-				tex = pST[1].t - pST[0].t;
-				if (tex == 0)
-					tex = 1.0f;
-				temp = -pST[0].t;
-				ratio = temp / tex;
-
-				wall = pXYZ[0].z - pXYZ[1].z;
-				temp = wall * ratio;
-				pXYZ[0].z -= temp;
-				pST[0].t = 0.0f;
-			}
-			if (pST[3].t < 0.0f)
-			{
-				float	tex, wall, ratio, temp;
-
-				tex = pST[2].t - pST[3].t;
-				if (tex == 0)
-					tex = 1.0f;
-				temp = -pST[3].t;
-				ratio = temp / tex;
-
-				wall = pXYZ[3].z - pXYZ[2].z;
-				temp = wall * ratio;
-				pXYZ[3].z -= temp;
-				pST[3].t = 0.0f;
-			}
-
-			pXYZ[1].z -= 16.0f;
-			pXYZ[2].z -= 16.0f;
-		}
-
-		pST[0].t += 1.0f / pDib->width;
-		pST[3].t += 1.0f / pDib->width;
-		pST[1].t -= 1.0f / pDib->width;
-		pST[2].t -= 1.0f / pDib->width;
-	}
-
-	if (pBGRA)
-	{
-		int	i;
-    float a, b;
-		int	distX, distY, distance;
-		long	lightScale;
-		long lo_end = FINENESS-shade_amount;
-
-		for (i = 0; i < 4; i++)
-		{
-			distX = pXYZ[i].x - player.x;
-			distY = pXYZ[i].y - player.y;
-
-			distance = DistanceGet(distX, distY);
-
-			if (shade_amount != 0)
-			{
-				a = pWall->separator.a;
-				b = pWall->separator.b;
-
-				if (side < 0)
-				{
-					a = -a;
-					b = -b;
-				}
-
-				lightScale = (long)(a * sun_vect.x +
-								b * sun_vect.y) >> LOG_FINENESS;
-
-				lightScale = (lightScale + FINENESS)>>1; // map to 0 to 1 range
-
-				lightScale = lo_end + ((lightScale * shade_amount)>>LOG_FINENESS);
-				
-				if (lightScale > FINENESS)
-					lightScale = FINENESS;
-				else if ( lightScale < 0)
-					lightScale = 0;
-			}
-			else
-				lightScale = FINENESS;
-
-			pWall->lightscale = lightScale;
-
-			if (gD3DDriverProfile.bFogEnable)
-				paletteIndex = GetLightPaletteIndex(FINENESS, light, lightScale, 0);
-			else
-				paletteIndex = GetLightPaletteIndex(distance, light, lightScale, 0);
-
-			if (light <= 127)
-				*flags |= D3DRENDER_NOAMBIENT;
-
-			pBGRA[i].r = pBGRA[i].g = pBGRA[i].b = paletteIndex * COLOR_AMBIENT / 64;
-			pBGRA[i].a = 255;
-		}
-	}
-
-	return 1;
-}
-
-void D3DRenderFloorExtract(BSPnode *pNode, PDIB pDib, custom_xyz *pXYZ, custom_st *pST,
-							custom_bgra *pBGRA)
-{
-	Sector	*pSector = pNode->u.leaf.sector;
-	int		count;
-	int		left, top;
-	int		paletteIndex;
-	float	oneOverC, inv128, inv64;
-	custom_xyz	intersectTop, intersectLeft;
-	long		lightscale;
-
-	left = top = 0;
-
-	inv128 = 1.0f / (128.0f * PETER_FUDGE);
-	inv64 = 1.0f / (64.0f * PETER_FUDGE);
-
-	// generate texture coordinates
-	if (pSector->sloped_floor)
-	{
-		left = pSector->sloped_floor->p0.x;
-		top = pSector->sloped_floor->p0.y;
-	}
-	else
-	{
-		for (count = 0; count < pNode->u.leaf.poly.npts; count++)
-		{
-			if (pNode->u.leaf.poly.p[count].x < left)
-				left = pNode->u.leaf.poly.p[count].x;
-		}
-		for (count = 0; count < pNode->u.leaf.poly.npts; count++)
-		{
-			if (pNode->u.leaf.poly.p[count].y < top)
-				top = pNode->u.leaf.poly.p[count].y;
-		}
-	}
-
-	if (pSector->sloped_floor)
-	{
-		oneOverC = 1.0f / pSector->sloped_floor->plane.c;
-	}
-
-	if (pXYZ)
-	{
-		for (count = 0; count < pNode->u.leaf.poly.npts; count++)
-		{
-			if (pSector->sloped_floor)
-			{
-				pXYZ[count].x = (float)pNode->u.leaf.poly.p[count].x;
-				pXYZ[count].y = (float)pNode->u.leaf.poly.p[count].y;
-				pXYZ[count].z = (-pSector->sloped_floor->plane.a * pXYZ[count].x -
-					pSector->sloped_floor->plane.b * pXYZ[count].y -
-					pSector->sloped_floor->plane.d) * oneOverC;
-			}
-			else
-			{
-				pXYZ[count].x = (float)pNode->u.leaf.poly.p[count].x;
-				pXYZ[count].y = (float)pNode->u.leaf.poly.p[count].y;
-				pXYZ[count].z = (float)pSector->floor_height;
-			}
-
-			if (pST)
-			{
-				custom_xyz	vectorU, vectorV, vector;
-				float		U, temp;
-
-				if (pSector->sloped_floor)
-				{
-					float	distance;
-
-					// calc distance from top line (vector u)
-					U = ((pXYZ[count].x - pSector->sloped_floor->p0.x) *
-						(pSector->sloped_floor->p1.x - pSector->sloped_floor->p0.x)) +
-						((pXYZ[count].z - pSector->sloped_floor->p0.z) *
-						(pSector->sloped_floor->p1.z - pSector->sloped_floor->p0.z)) +
-						((pXYZ[count].y - pSector->sloped_floor->p0.y) *
-						(pSector->sloped_floor->p1.y - pSector->sloped_floor->p0.y));
-					temp = ((pSector->sloped_floor->p1.x - pSector->sloped_floor->p0.x) *
-						(pSector->sloped_floor->p1.x - pSector->sloped_floor->p0.x)) +
-						((pSector->sloped_floor->p1.z - pSector->sloped_floor->p0.z) *
-						(pSector->sloped_floor->p1.z - pSector->sloped_floor->p0.z)) +
-						((pSector->sloped_floor->p1.y - pSector->sloped_floor->p0.y) *
-						(pSector->sloped_floor->p1.y - pSector->sloped_floor->p0.y));
-
-					if (temp == 0)
-						temp = 1.0f;
-
-					U /= temp;
-
-					intersectTop.x = pSector->sloped_floor->p0.x +
-						U * (pSector->sloped_floor->p1.x - pSector->sloped_floor->p0.x);
-					intersectTop.z = pSector->sloped_floor->p0.z +
-						U * (pSector->sloped_floor->p1.z - pSector->sloped_floor->p0.z);
-					intersectTop.y = pSector->sloped_floor->p0.y +
-						U * (pSector->sloped_floor->p1.y - pSector->sloped_floor->p0.y);
-
-					pST[count].s = (float)sqrt((pXYZ[count].x - intersectTop.x) *
-									(pXYZ[count].x - intersectTop.x) +
-									(pXYZ[count].z - intersectTop.z) *
-									(pXYZ[count].z - intersectTop.z) +
-									(pXYZ[count].y - intersectTop.y) *
-									(pXYZ[count].y - intersectTop.y));
-
-					// calc distance from left line (vector v)
-					U = ((pXYZ[count].x - pSector->sloped_floor->p0.x) *
-						(pSector->sloped_floor->p2.x - pSector->sloped_floor->p0.x)) +
-						((pXYZ[count].z - pSector->sloped_floor->p0.z) *
-						(pSector->sloped_floor->p2.z - pSector->sloped_floor->p0.z)) +
-						((pXYZ[count].y - pSector->sloped_floor->p0.y) *
-						(pSector->sloped_floor->p2.y - pSector->sloped_floor->p0.y));
-					temp = ((pSector->sloped_floor->p2.x - pSector->sloped_floor->p0.x) *
-						(pSector->sloped_floor->p2.x - pSector->sloped_floor->p0.x)) +
-						((pSector->sloped_floor->p2.z - pSector->sloped_floor->p0.z) *
-						(pSector->sloped_floor->p2.z - pSector->sloped_floor->p0.z)) +
-						((pSector->sloped_floor->p2.y - pSector->sloped_floor->p0.y) *
-						(pSector->sloped_floor->p2.y - pSector->sloped_floor->p0.y));
-
-					if (temp == 0)
-						temp = 1.0f;
-
-					U /= temp;
-
-					intersectLeft.x = pSector->sloped_floor->p0.x +
-						U * (pSector->sloped_floor->p2.x - pSector->sloped_floor->p0.x);
-					intersectLeft.z = pSector->sloped_floor->p0.z +
-						U * (pSector->sloped_floor->p2.z - pSector->sloped_floor->p0.z);
-					intersectLeft.y = pSector->sloped_floor->p0.y +
-						U * (pSector->sloped_floor->p2.y - pSector->sloped_floor->p0.y);
-
-					pST[count].t = (float)sqrt((pXYZ[count].x - intersectLeft.x) *
-									(pXYZ[count].x - intersectLeft.x) +
-									(pXYZ[count].z - intersectLeft.z) *
-									(pXYZ[count].z - intersectLeft.z) +
-									(pXYZ[count].y - intersectLeft.y) *
-									(pXYZ[count].y - intersectLeft.y));
-
-					pST[count].s += pSector->ty / 2.0f;
-					pST[count].t += pSector->tx / 2.0f;
-
-					vectorU.x = pSector->sloped_floor->p1.x - pSector->sloped_floor->p0.x;
-					vectorU.z = pSector->sloped_floor->p1.z - pSector->sloped_floor->p0.z;
-					vectorU.y = pSector->sloped_floor->p1.y - pSector->sloped_floor->p0.y;
-
-					distance = (float)sqrt((vectorU.x * vectorU.x) +
-						(vectorU.y * vectorU.y));
-
-					if (distance == 0)
-						distance = 1.0f;
-
-					vectorU.x /= distance;
-					vectorU.z /= distance;
-					vectorU.y /= distance;
-
-					vectorV.x = pSector->sloped_floor->p2.x - pSector->sloped_floor->p0.x;
-					vectorV.z = pSector->sloped_floor->p2.z - pSector->sloped_floor->p0.z;
-					vectorV.y = pSector->sloped_floor->p2.y - pSector->sloped_floor->p0.y;
-
-					distance = (float)sqrt((vectorV.x * vectorV.x) +
-						(vectorV.y * vectorV.y));
-
-					if (distance == 0)
-						distance = 1.0f;
-
-					vectorV.x /= distance;
-					vectorV.z /= distance;
-					vectorV.y /= distance;
-
-					vector.x = pXYZ[count].x - pSector->sloped_floor->p0.x;
-					vector.y = pXYZ[count].y - pSector->sloped_floor->p0.y;
-
-					distance = (float)sqrt((vector.x * vector.x) +
-						(vector.y * vector.y));
-
-					if (distance == 0)
-						distance = 1.0f;
-
-					vector.x /= distance;
-					vector.y /= distance;
-
-					if (((vector.x * vectorU.x) +
-						(vector.y * vectorU.y)) <= 0)
-						pST[count].t = -pST[count].t;
-
-					if (((vector.x * vectorV.x) +
-						(vector.y * vectorV.y)) > 0)
-						pST[count].s = -pST[count].s;
-				}
-				else
-				{
-					pST[count].s = abs(pNode->u.leaf.poly.p[count].y - top) - pSector->ty;
-					pST[count].t = abs(pNode->u.leaf.poly.p[count].x - left) - pSector->tx;
-				}
-
-				if (pSector->animate != NULL && pSector->animate->animation == ANIMATE_SCROLL)
-				{
-					if (pSector->flags & SF_SCROLL_FLOOR)
-					{
-						pST[count].s -= pSector->animate->u.scroll.yoffset;
-						pST[count].t += pSector->animate->u.scroll.xoffset;
-					}
-				}
-
-        pST[count].s *= inv64;
-        pST[count].t *= inv64;
-			}
-		}
-	}
-
-	if (pBGRA)
-	{
-		int	distX, distY, distance;
-
-		for (count = 0; count < pNode->u.leaf.poly.npts; count++)
-		{
-			distX = pXYZ[count].x - player.x;
-			distY = pXYZ[count].y - player.y;
-
-			lightscale = FINENESS;
-
-			if (shade_amount != 0)
-			{
-				// floor is sloped, and it's marked as being steep enough to be eligible for
-				// directional lighting
-				if ((pNode->u.leaf.sector->sloped_floor != NULL) &&
-					(pNode->u.leaf.sector->sloped_floor->flags & SLF_DIRECTIONAL))
-				{
-					long lo_end = FINENESS-shade_amount;
-
-					// light scale is based on dot product of surface normal and sun vector
-					lightscale = (long)(pNode->u.leaf.sector->sloped_floor->plane.a * sun_vect.x +
-						pNode->u.leaf.sector->sloped_floor->plane.b * sun_vect.y +
-						pNode->u.leaf.sector->sloped_floor->plane.c * sun_vect.z)>>LOG_FINENESS;
-
-					lightscale = (lightscale + FINENESS)>>1; // map to 0 to 1 range
-
-					lightscale = lo_end + ((lightscale * shade_amount)>>LOG_FINENESS);
-
-					if (lightscale > FINENESS)
-						lightscale = FINENESS;
-					else if ( lightscale < 0)
-						lightscale = 0;
-
-					pNode->u.leaf.sector->sloped_floor->lightscale = lightscale;
-				}
-			}
-			else
-			{ // normal light intensity
-				if (pNode->u.leaf.sector->sloped_floor != NULL)
-					pNode->u.leaf.sector->sloped_floor->lightscale = FINENESS;
-			}
-
-			distance = DistanceGet(distX, distY);
-
-			if (gD3DDriverProfile.bFogEnable)
-				paletteIndex = GetLightPaletteIndex(FINENESS, pSector->light, lightscale, 0);
-			else
-				paletteIndex = GetLightPaletteIndex(distance, pSector->light, lightscale, 0);
-
-			pBGRA[count].r = pBGRA[count].g = pBGRA[count].b = paletteIndex * COLOR_AMBIENT / 64;
-			pBGRA[count].a = 255;
-		}
-	}
-}
-
-void D3DRenderCeilingExtract(BSPnode *pNode, PDIB pDib, custom_xyz *pXYZ, custom_st *pST,
-							custom_bgra *pBGRA)
-{
-	Sector		*pSector = pNode->u.leaf.sector;
-	int			count;
-	int			left, top;
-	int			paletteIndex;
-	float		oneOverC, inv128, inv64;
-	custom_xyz	intersectTop, intersectLeft;
-	long		lightscale;
-
-	left = top = 0;
-
-	inv128 = 1.0f / (128.0f * PETER_FUDGE);
-	inv64 = 1.0f / (64.0f * PETER_FUDGE);
-
-	// generate texture coordinates
-	for (count = 0; count < pNode->u.leaf.poly.npts; count++)
-	{
-		if (pNode->u.leaf.poly.p[count].x < left)
-			left = pNode->u.leaf.poly.p[count].x;
-	}
-	for (count = 0; count < pNode->u.leaf.poly.npts; count++)
-	{
-		if (pNode->u.leaf.poly.p[count].y < top)
-			top = pNode->u.leaf.poly.p[count].y;
-	}
-
-	if (pSector->sloped_ceiling)
-	{
-		// extract plane normal
-		oneOverC = 1.0f / pSector->sloped_ceiling->plane.c;
-	}
-
-	if (pXYZ)
-	{
-		for (count = 0; count < pNode->u.leaf.poly.npts; count++)
-		{
-			if (pSector->sloped_ceiling)
-			{
-				pXYZ[count].x = (float)pNode->u.leaf.poly.p[count].x;
-				pXYZ[count].y = (float)pNode->u.leaf.poly.p[count].y;
-				pXYZ[count].z = (-pSector->sloped_ceiling->plane.a * pXYZ[count].x -
-					pSector->sloped_ceiling->plane.b * pXYZ[count].y -
-					pSector->sloped_ceiling->plane.d) * oneOverC;
-			}
-			else
-			{
-				pXYZ[count].x = (float)pNode->u.leaf.poly.p[count].x;
-				pXYZ[count].y = (float)pNode->u.leaf.poly.p[count].y;
-				pXYZ[count].z = (float)pSector->ceiling_height;
-			}
-
-			if (pST)
-			{
-				custom_xyz	vectorU, vectorV, vector;
-				float		U, temp;
-
-				if (pSector->sloped_ceiling)
-				{
-					float	distance;
-
-					// calc distance from top line (vector u)
-					U = ((pXYZ[count].x - pSector->sloped_ceiling->p0.x) *
-						(pSector->sloped_ceiling->p1.x - pSector->sloped_ceiling->p0.x)) +
-						((pXYZ[count].z - pSector->sloped_ceiling->p0.z) *
-						(pSector->sloped_ceiling->p1.z - pSector->sloped_ceiling->p0.z)) +
-						((pXYZ[count].y - pSector->sloped_ceiling->p0.y) *
-						(pSector->sloped_ceiling->p1.y - pSector->sloped_ceiling->p0.y));
-					temp = ((pSector->sloped_ceiling->p1.x - pSector->sloped_ceiling->p0.x) *
-						(pSector->sloped_ceiling->p1.x - pSector->sloped_ceiling->p0.x)) +
-						((pSector->sloped_ceiling->p1.z - pSector->sloped_ceiling->p0.z) *
-						(pSector->sloped_ceiling->p1.z - pSector->sloped_ceiling->p0.z)) +
-						((pSector->sloped_ceiling->p1.y - pSector->sloped_ceiling->p0.y) *
-						(pSector->sloped_ceiling->p1.y - pSector->sloped_ceiling->p0.y));
-
-					if (temp == 0)
-						temp = 1.0f;
-
-					U /= temp;
-
-					intersectTop.x = pSector->sloped_ceiling->p0.x +
-						U * (pSector->sloped_ceiling->p1.x - pSector->sloped_ceiling->p0.x);
-					intersectTop.z = pSector->sloped_ceiling->p0.z +
-						U * (pSector->sloped_ceiling->p1.z - pSector->sloped_ceiling->p0.z);
-					intersectTop.y = pSector->sloped_ceiling->p0.y +
-						U * (pSector->sloped_ceiling->p1.y - pSector->sloped_ceiling->p0.y);
-
-					pST[count].s = (float)sqrt((pXYZ[count].x - intersectTop.x) *
-									(pXYZ[count].x - intersectTop.x) +
-									(pXYZ[count].z - intersectTop.z) *
-									(pXYZ[count].z - intersectTop.z) +
-									(pXYZ[count].y - intersectTop.y) *
-									(pXYZ[count].y - intersectTop.y));
-
-					// calc distance from left line (vector v)
-					U = ((pXYZ[count].x - pSector->sloped_ceiling->p0.x) *
-						(pSector->sloped_ceiling->p2.x - pSector->sloped_ceiling->p0.x)) +
-						((pXYZ[count].z - pSector->sloped_ceiling->p0.z) *
-						(pSector->sloped_ceiling->p2.z - pSector->sloped_ceiling->p0.z)) +
-						((pXYZ[count].y - pSector->sloped_ceiling->p0.y) *
-						(pSector->sloped_ceiling->p2.y - pSector->sloped_ceiling->p0.y));
-					temp = ((pSector->sloped_ceiling->p2.x - pSector->sloped_ceiling->p0.x) *
-						(pSector->sloped_ceiling->p2.x - pSector->sloped_ceiling->p0.x)) +
-						((pSector->sloped_ceiling->p2.z - pSector->sloped_ceiling->p0.z) *
-						(pSector->sloped_ceiling->p2.z - pSector->sloped_ceiling->p0.z)) +
-						((pSector->sloped_ceiling->p2.y - pSector->sloped_ceiling->p0.y) *
-						(pSector->sloped_ceiling->p2.y - pSector->sloped_ceiling->p0.y));
-
-					if (temp == 0)
-						temp = 1.0f;
-
-					U /= temp;
-
-					intersectLeft.x = pSector->sloped_ceiling->p0.x +
-						U * (pSector->sloped_ceiling->p2.x - pSector->sloped_ceiling->p0.x);
-					intersectLeft.z = pSector->sloped_ceiling->p0.z +
-						U * (pSector->sloped_ceiling->p2.z - pSector->sloped_ceiling->p0.z);
-					intersectLeft.y = pSector->sloped_ceiling->p0.y +
-						U * (pSector->sloped_ceiling->p2.y - pSector->sloped_ceiling->p0.y);
-
-					pST[count].t = (float)sqrt((pXYZ[count].x - intersectLeft.x) *
-									(pXYZ[count].x - intersectLeft.x) +
-									(pXYZ[count].z - intersectLeft.z) *
-									(pXYZ[count].z - intersectLeft.z) +
-									(pXYZ[count].y - intersectLeft.y) *
-									(pXYZ[count].y - intersectLeft.y));
-
-					vectorU.x = pSector->sloped_ceiling->p1.x - pSector->sloped_ceiling->p0.x;
-					vectorU.z = pSector->sloped_ceiling->p1.z - pSector->sloped_ceiling->p0.z;
-					vectorU.y = pSector->sloped_ceiling->p1.y - pSector->sloped_ceiling->p0.y;
-
-					distance = (float)sqrt((vectorU.x * vectorU.x) +
-						(vectorU.y * vectorU.y));
-
-					if (distance == 0)
-						distance = 1.0f;
-
-					vectorU.x /= distance;
-					vectorU.z /= distance;
-					vectorU.y /= distance;
-
-					vectorV.x = pSector->sloped_ceiling->p2.x - pSector->sloped_ceiling->p0.x;
-					vectorV.z = pSector->sloped_ceiling->p2.z - pSector->sloped_ceiling->p0.z;
-					vectorV.y = pSector->sloped_ceiling->p2.y - pSector->sloped_ceiling->p0.y;
-
-					distance = (float)sqrt((vectorV.x * vectorV.x) +
-						(vectorV.y * vectorV.y));
-
-					if (distance == 0)
-						distance = 1.0f;
-
-					vectorV.x /= distance;
-					vectorV.z /= distance;
-					vectorV.y /= distance;
-
-					vector.x = pXYZ[count].x - pSector->sloped_ceiling->p0.x;
-					vector.y = pXYZ[count].y - pSector->sloped_ceiling->p0.y;
-
-					distance = (float)sqrt((vector.x * vector.x) +
-						(vector.y * vector.y));
-
-					if (distance == 0)
-						distance = 1.0f;
-
-					vector.x /= distance;
-					vector.y /= distance;
-
-					if (((vector.x * vectorU.x) +
-						(vector.y * vectorU.y)) < 0)
-						pST[count].t = -pST[count].t;
-
-					if (((vector.x * vectorV.x) +
-						(vector.y * vectorV.y)) > 0)
-						pST[count].s = -pST[count].s;
-
-					pST[count].s -= pSector->ty / 2.0f;
-					pST[count].t -= pSector->tx / 2.0f;
-				}
-				else
-				{
-					pST[count].s = abs(pNode->u.leaf.poly.p[count].y - top) - pSector->ty;
-					pST[count].t = abs(pNode->u.leaf.poly.p[count].x - left) - pSector->tx;
-				}
-
-				if (pSector->animate != NULL && pSector->animate->animation == ANIMATE_SCROLL)
-				{
-					if (pSector->flags & SF_SCROLL_CEILING)
-					{
-						pST[count].s -= pSector->animate->u.scroll.yoffset;
-						pST[count].t += pSector->animate->u.scroll.xoffset;
-					}
-				}
-
-            pST[count].s *= inv64;
-            pST[count].t *= inv64;
-			}
-		}
-	}
-
-	if (pBGRA)
-	{
-		int	distX, distY, distance;
-
-		for (count = 0; count < pNode->u.leaf.poly.npts; count++)
-		{
-			distX = pXYZ[count].x - player.x;
-			distY = pXYZ[count].y - player.y;
-
-			lightscale = FINENESS;
-
-			if (shade_amount != 0)
-			{
-				if ((pNode->u.leaf.sector->sloped_ceiling != NULL) &&
-					(pNode->u.leaf.sector->sloped_ceiling->flags & SLF_DIRECTIONAL))
-				{
-					long lo_end = FINENESS-shade_amount;
-
-					// light scale is based on dot product of surface normal and sun vector
-					lightscale = (long)(pNode->u.leaf.sector->sloped_ceiling->plane.a * sun_vect.x +
-						pNode->u.leaf.sector->sloped_ceiling->plane.b * sun_vect.y +
-						pNode->u.leaf.sector->sloped_ceiling->plane.a * sun_vect.z)>>LOG_FINENESS;
-
-					lightscale = (lightscale + FINENESS)>>1; // map to 0 to 1 range
-
-					lightscale = lo_end + ((lightscale * shade_amount)>>LOG_FINENESS);
-
-					if (lightscale > FINENESS)
-						lightscale = FINENESS;
-					else if ( lightscale < 0)
-						lightscale = 0;
-
-					pNode->u.leaf.sector->sloped_ceiling->lightscale = lightscale;
-				}
-			}
-			else
-			{
-				if (pNode->u.leaf.sector->sloped_ceiling != NULL)
-					pNode->u.leaf.sector->sloped_ceiling->lightscale = FINENESS;
-			}
-
-			distance = DistanceGet(distX, distY);
-
-			if (gD3DDriverProfile.bFogEnable)
-				paletteIndex = GetLightPaletteIndex(FINENESS, pSector->light, lightscale, 0);
-			else
-				paletteIndex = GetLightPaletteIndex(distance, pSector->light, lightscale, 0);
-
-			pBGRA[count].r = pBGRA[count].g = pBGRA[count].b = paletteIndex * COLOR_AMBIENT / 64;
-			pBGRA[count].a = 255;
-		}
-	}
-}
-
 
 // new render stuff
 void D3DRenderPoolInit(d3d_render_pool_new *pPool, int size, int packetSize)
